// src/graphql/types.ts
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: any;
  /** An ISO-8601 encoded UTC date string. */
  DateTime: any;
  /** A string containing HTML code. */
  HTML: any;
  /** A Git object ID. */
  GitObjectID: any;
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: any;
  /** Git SSH string */
  GitSSHRemote: any;
  /** An ISO-8601 encoded date string. */
  Date: any;
  /** A valid x509 certificate string */
  X509Certificate: any;
};
// ...
export type Actor = {
  /** A URL pointing to the actor's public avatar. */
  avatarUrl: Scalars["URI"];
  /** The username of the actor. */
  login: Scalars["String"];
  /** The HTTP path for this actor. */
  resourcePath: Scalars["URI"];
  /** The HTTP URL for this actor. */
  url: Scalars["URI"];
};
// ...
export type Assignable = {
  /** A list of Users assigned to this object. */
  assignees: UserConnection;
};
// ...
export type Closable = {
  /** `true` if the object is closed (definition of closed may depend on type) */
  closed: Scalars["Boolean"];
  /** Identifies the date and time when the object was closed. */
  closedAt?: Maybe<Scalars["DateTime"]>;
};
// ...
export type Comment = {
  /** The actor who authored the comment. */
  author?: Maybe<Actor>;
  /** Author's association with the subject of the comment. */
  authorAssociation: CommentAuthorAssociation;
  /** The body as Markdown. */
  body: Scalars["String"];
  /** The body rendered to HTML. */
  bodyHTML: Scalars["HTML"];
  /** The body rendered to text. */
  bodyText: Scalars["String"];
  /** Identifies the date and time when the object was created. */
  createdAt: Scalars["DateTime"];
  /** Check if this comment was created via an email reply. */
  createdViaEmail: Scalars["Boolean"];
  /** The actor who edited the comment. */
  editor?: Maybe<Actor>;
  id: Scalars["ID"];
  /** Check if this comment was edited and includes an edit with the creation data */
  includesCreatedEdit: Scalars["Boolean"];
  /** The moment the editor made the last edit */
  lastEditedAt?: Maybe<Scalars["DateTime"]>;
  /** Identifies when the comment was published at. */
  publishedAt?: Maybe<Scalars["DateTime"]>;
  /** Identifies the date and time when the object was last updated. */
  updatedAt: Scalars["DateTime"];
  /** A list of edits to this content. */
  userContentEdits?: Maybe<UserContentEditConnection>;
  /** Did the viewer author this comment. */
  viewerDidAuthor: Scalars["Boolean"];
};

/** Represents a comment. */
export type CommentUserContentEditsArgs = {
  after?: Maybe<Scalars["String"]>;
  before?: Maybe<Scalars["String"]>;
  first?: Maybe<Scalars["Int"]>;
  last?: Maybe<Scalars["Int"]>;
};

/** A comment author association with repository. */
export enum CommentAuthorAssociation {
  /** Author is a member of the organization that owns the repository. */
  Member = "MEMBER",
  /** Author is the owner of the repository. */
  Owner = "OWNER",
  /** Author has been invited to collaborate on the repository. */
  Collaborator = "COLLABORATOR",
  /** Author has previously committed to the repository. */
  Contributor = "CONTRIBUTOR",
  /** Author has not previously committed to the repository. */
  FirstTimeContributor = "FIRST_TIME_CONTRIBUTOR",
  /** Author has not previously committed to GitHub. */
  FirstTimer = "FIRST_TIMER",
  /** Author has no association with the repository. */
  None = "NONE"
}
// ...
export type Issue = Node &
  Assignable &
  Closable &
  Comment &
  Updatable &
  UpdatableComment &
  Labelable &
  Lockable &
  Reactable &
  RepositoryNode &
  Subscribable &
  UniformResourceLocatable & {
    __typename?: "Issue";
    /** Reason that the conversation was locked. */
    activeLockReason?: Maybe<LockReason>;
    /** A list of Users assigned to this object. */
    assignees: UserConnection;
    /** The actor who authored the comment. */
    author?: Maybe<Actor>;
    /** Author's association with the subject of the comment. */
    authorAssociation: CommentAuthorAssociation;
    /** Identifies the body of the issue. */
    body: Scalars["String"];
    /** Identifies the body of the issue rendered to HTML. */
    bodyHTML: Scalars["HTML"];
    /** Identifies the body of the issue rendered to text. */
    bodyText: Scalars["String"];
    /** `true` if the object is closed (definition of closed may depend on type) */
    closed: Scalars["Boolean"];
    /** Identifies the date and time when the object was closed. */
    closedAt?: Maybe<Scalars["DateTime"]>;
    /** A list of comments associated with the Issue. */
    comments: IssueCommentConnection;
    /** Identifies the date and time when the object was created. */
    createdAt: Scalars["DateTime"];
    /** Check if this comment was created via an email reply. */
    createdViaEmail: Scalars["Boolean"];
    /** Identifies the primary key from the database. */
    databaseId?: Maybe<Scalars["Int"]>;
    /** The actor who edited the comment. */
    editor?: Maybe<Actor>;
    id: Scalars["ID"];
    /** Check if this comment was edited and includes an edit with the creation data */
    includesCreatedEdit: Scalars["Boolean"];
    /** A list of labels associated with the object. */
    labels?: Maybe<LabelConnection>;
    /** The moment the editor made the last edit */
    lastEditedAt?: Maybe<Scalars["DateTime"]>;
    /** `true` if the object is locked */
    locked: Scalars["Boolean"];
    /** Identifies the milestone associated with the issue. */
    milestone?: Maybe<Milestone>;
    /** Identifies the issue number. */
    number: Scalars["Int"];
    /** A list of Users that are participating in the Issue conversation. */
    participants: UserConnection;
    /** List of project cards associated with this issue. */
    projectCards: ProjectCardConnection;
    /** Identifies when the comment was published at. */
    publishedAt?: Maybe<Scalars["DateTime"]>;
    /** A list of reactions grouped by content left on the subject. */
    reactionGroups?: Maybe<Array<ReactionGroup>>;
    /** A list of Reactions left on the Issue. */
    reactions: ReactionConnection;
    /** The repository associated with this node. */
    repository: Repository;
    /** The HTTP path for this issue */
    resourcePath: Scalars["URI"];
    /** Identifies the state of the issue. */
    state: IssueState;
    /** A list of events, comments, commits, etc. associated with the issue. */
    timeline: IssueTimelineConnection;
    /** A list of events, comments, commits, etc. associated with the issue. */
    timelineItems: IssueTimelineItemsConnection;
    /** Identifies the issue title. */
    title: Scalars["String"];
    /** Identifies the date and time when the object was last updated. */
    updatedAt: Scalars["DateTime"];
    /** The HTTP URL for this issue */
    url: Scalars["URI"];
    /** A list of edits to this content. */
    userContentEdits?: Maybe<UserContentEditConnection>;
    /** Can user react to this subject */
    viewerCanReact: Scalars["Boolean"];
    /** Check if the viewer is able to change their subscription status for the repository. */
    viewerCanSubscribe: Scalars["Boolean"];
    /** Check if the current viewer can update this object. */
    viewerCanUpdate: Scalars["Boolean"];
    /** Reasons why the current viewer can not update this comment. */
    viewerCannotUpdateReasons: Array<CommentCannotUpdateReason>;
    /** Did the viewer author this comment. */
    viewerDidAuthor: Scalars["Boolean"];
    /** Identifies if the viewer is watching, not watching, or ignoring the subscribable entity. */
    viewerSubscription?: Maybe<SubscriptionState>;
  };
// ...
export type IssueCommentConnection = {
  __typename?: "IssueCommentConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<IssueCommentEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<IssueComment>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};
// ...
export type Labelable = {
  /** A list of labels associated with the object. */
  labels?: Maybe<LabelConnection>;
};

/** An object that can have labels assigned to it. */
export type LabelableLabelsArgs = {
  after?: Maybe<Scalars["String"]>;
  before?: Maybe<Scalars["String"]>;
  first?: Maybe<Scalars["Int"]>;
  last?: Maybe<Scalars["Int"]>;
};

/** The connection type for Label. */
export type LabelConnection = {
  __typename?: "LabelConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<LabelEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<Label>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};
// ...
export type Lockable = {
  /** Reason that the conversation was locked. */
  activeLockReason?: Maybe<LockReason>;
  /** `true` if the object is locked */
  locked: Scalars["Boolean"];
};
// ...
export enum LockReason {
  /** The issue or pull request was locked because the conversation was off-topic. */
  OffTopic = "OFF_TOPIC",
  /** The issue or pull request was locked because the conversation was too heated. */
  TooHeated = "TOO_HEATED",
  /** The issue or pull request was locked because the conversation was resolved. */
  Resolved = "RESOLVED",
  /** The issue or pull request was locked because the conversation was spam. */
  Spam = "SPAM"
}
// ...
export type Node = {
  /** ID of the object. */
  id: Scalars["ID"];
};
// ...
export type ProjectCardConnection = {
  __typename?: "ProjectCardConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<ProjectCardEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<ProjectCard>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};
// ...
export type Reactable = {
  /** Identifies the primary key from the database. */
  databaseId?: Maybe<Scalars["Int"]>;
  id: Scalars["ID"];
  /** A list of reactions grouped by content left on the subject. */
  reactionGroups?: Maybe<Array<ReactionGroup>>;
  /** A list of Reactions left on the Issue. */
  reactions: ReactionConnection;
  /** Can user react to this subject */
  viewerCanReact: Scalars["Boolean"];
};
// ...
export type ReactionGroup = {
  __typename?: "ReactionGroup";
  /** Identifies the emoji reaction. */
  content: ReactionContent;
  /** Identifies when the reaction was created. */
  createdAt?: Maybe<Scalars["DateTime"]>;
  /** The subject that was reacted to. */
  subject: Reactable;
  /** Users who have reacted to the reaction subject with the emotion represented by this reaction group */
  users: ReactingUserConnection;
  /** Whether or not the authenticated user has left a reaction on the subject. */
  viewerHasReacted: Scalars["Boolean"];
};
// ...
export type RepositoryNode = {
  /** The repository associated with this node. */
  repository: Repository;
};
// ...
export enum SearchType {
  /** Returns results matching issues in repositories. */
  Issue = "ISSUE",
  /** Returns results matching repositories. */
  Repository = "REPOSITORY",
  /** Returns results matching users and organizations on GitHub. */
  User = "USER"
}
// ...
export type Subscribable = {
  id: Scalars["ID"];
  /** Check if the viewer is able to change their subscription status for the repository. */
  viewerCanSubscribe: Scalars["Boolean"];
  /** Identifies if the viewer is watching, not watching, or ignoring the subscribable entity. */
  viewerSubscription?: Maybe<SubscriptionState>;
};
// ...
export type UniformResourceLocatable = {
  /** The HTML path to this resource. */
  resourcePath: Scalars["URI"];
  /** The URL to this resource. */
  url: Scalars["URI"];
};
// ...
export type Updatable = {
  /** Check if the current viewer can update this object. */
  viewerCanUpdate: Scalars["Boolean"];
};

/** Comments that can be updated. */
export type UpdatableComment = {
  /** Reasons why the current viewer can not update this comment. */
  viewerCannotUpdateReasons: Array<CommentCannotUpdateReason>;
};
// ...
export type UserConnection = {
  __typename?: "UserConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<UserEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<User>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};